requirements.txt
evidently==0.5.1
Flask==3.1.0
flask-apscheduler==1.13.1
Flask-Caching==2.3.0
flask-cors==5.0.0
Gunicorn
hopsworks[python]==4.1.*
huggingface_hub==0.29.3
numpy<2.0.0
pandas==1.5.3
pydantic==2.11.1
pydantic-settings==2.7.1
python-dotenv==1.1.0
tensorflow==2.18.0
yfinance==0.2.54
git+https://github.com/marcus-24/Stock-Predictor-Feature-Store